Move i3pyblocks.__main__ to i3pyblocks.cli, improve tests
